pyproject.toml
[project]
name = "pooolify"
version = "0.1.0"
description = "pooolify: Code-centric Python framework for multi-agent LLM orchestration with FastAPI + PostgreSQL persistence"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "pooolify maintainers" }]
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.110",
  "pydantic>=2.6",
  "uvicorn>=0.30",
  "httpx>=0.27",
  "openai>=1.40.0",
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/Pooolingforest/pooolify"

[tool.uv]
dev-dependencies = [
  "pytest",
  "anyio",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"


